+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+* feat: improve model factory `state` method by @lptn in https://github.com/nunomaduro/larastan/pull/1219
 
 ## [2.1.4] - 2022-03-30