+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Improvements
+* feat: Detect actual date class in Date facade, model attributes, now and today helpers by @bastien-phi in https://github.com/nunomaduro/larastan/pull/1142
+
 ### Fixes
 
 * fix: Change QueryBuilder::newQuery() @return from `$this` to `static`